feat(ai) : AI 가 추천해준 방식으로 리팩토링
